+++ jgrapht-capi/src/main/java/org/jgrapht/capi/impl/CycleApi.java
@@ -36,11 +36,11 @@ public class CycleApi {
 			+ "cycles_eulerian_exec_hierholzer", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeHierholzer(IsolateThread thread, ObjectHandle graphHandle, CIntPointer isEulerianRes,
 			WordPointer eulerianCycleRes) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
 
-		HierholzerEulerianCycle<Long, Long> alg = new HierholzerEulerianCycle<>();
+		HierholzerEulerianCycle<Integer, Integer> alg = new HierholzerEulerianCycle<>();
 
-		GraphPath<Long, Long> eulerianCycle = null;
+		GraphPath<Integer, Integer> eulerianCycle = null;
 		try {
 			eulerianCycle = alg.getEulerianCycle(g);
 		} catch (IllegalArgumentException e) {
@@ -65,10 +65,10 @@ public class CycleApi {
 	@CEntryPoint(name = Constants.LIB_PREFIX
 			+ "cycles_chinese_postman_exec_edmonds_johnson", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeChinesePostman(IsolateThread thread, ObjectHandle graphHandle, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
 
-		ChinesePostman<Long, Long> alg = new ChinesePostman<>();
-		GraphPath<Long, Long> path = alg.getCPPSolution(g);
+		ChinesePostman<Integer, Integer> alg = new ChinesePostman<>();
+		GraphPath<Integer, Integer> path = alg.getCPPSolution(g);
 
 		if (res.isNonNull()) {
 			res.write(globalHandles.create(path));
@@ -79,8 +79,8 @@ public class CycleApi {
 	@CEntryPoint(name = Constants.LIB_PREFIX
 			+ "cycles_simple_enumeration_exec_tarjan", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeTarjan(IsolateThread thread, ObjectHandle graphHandle, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
-		DirectedSimpleCycles<Long, Long> alg = new TarjanSimpleCycles<>(g);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
+		DirectedSimpleCycles<Integer, Integer> alg = new TarjanSimpleCycles<>(g);
 		if (res.isNonNull()) {
 			res.write(globalHandles.create(alg.findSimpleCycles().iterator()));
 		}
@@ -90,8 +90,8 @@ public class CycleApi {
 	@CEntryPoint(name = Constants.LIB_PREFIX
 			+ "cycles_simple_enumeration_exec_tiernan", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeTierman(IsolateThread thread, ObjectHandle graphHandle, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
-		DirectedSimpleCycles<Long, Long> alg = new TiernanSimpleCycles<>(g);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
+		DirectedSimpleCycles<Integer, Integer> alg = new TiernanSimpleCycles<>(g);
 		if (res.isNonNull()) {
 			res.write(globalHandles.create(alg.findSimpleCycles().iterator()));
 		}
@@ -101,8 +101,8 @@ public class CycleApi {
 	@CEntryPoint(name = Constants.LIB_PREFIX
 			+ "cycles_simple_enumeration_exec_szwarcfiter_lauer", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeSzwarcfiterLauer(IsolateThread thread, ObjectHandle graphHandle, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
-		DirectedSimpleCycles<Long, Long> alg = new SzwarcfiterLauerSimpleCycles<>(g);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
+		DirectedSimpleCycles<Integer, Integer> alg = new SzwarcfiterLauerSimpleCycles<>(g);
 		if (res.isNonNull()) {
 			res.write(globalHandles.create(alg.findSimpleCycles().iterator()));
 		}
@@ -112,8 +112,8 @@ public class CycleApi {
 	@CEntryPoint(name = Constants.LIB_PREFIX
 			+ "cycles_simple_enumeration_exec_johnson", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeJohnson(IsolateThread thread, ObjectHandle graphHandle, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
-		DirectedSimpleCycles<Long, Long> alg = new JohnsonSimpleCycles<>(g);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
+		DirectedSimpleCycles<Integer, Integer> alg = new JohnsonSimpleCycles<>(g);
 		if (res.isNonNull()) {
 			res.write(globalHandles.create(alg.findSimpleCycles().iterator()));
 		}
@@ -123,8 +123,8 @@ public class CycleApi {
 	@CEntryPoint(name = Constants.LIB_PREFIX
 			+ "cycles_simple_enumeration_exec_hawick_james", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeHawickJames(IsolateThread thread, ObjectHandle graphHandle, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
-		DirectedSimpleCycles<Long, Long> alg = new HawickJamesSimpleCycles<>(g);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
+		DirectedSimpleCycles<Integer, Integer> alg = new HawickJamesSimpleCycles<>(g);
 		if (res.isNonNull()) {
 			res.write(globalHandles.create(alg.findSimpleCycles().iterator()));
 		}
@@ -135,12 +135,12 @@ public class CycleApi {
 			+ "cycles_fundamental_basis_exec_queue_bfs", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeQueueBFSFundamental(IsolateThread thread, ObjectHandle graphHandle,
 			CDoublePointer weightRes, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
 
-		CycleBasisAlgorithm<Long, Long> alg = new QueueBFSFundamentalCycleBasis<>(g);
-		CycleBasis<Long, Long> cycleBasis = alg.getCycleBasis();
+		CycleBasisAlgorithm<Integer, Integer> alg = new QueueBFSFundamentalCycleBasis<>(g);
+		CycleBasis<Integer, Integer> cycleBasis = alg.getCycleBasis();
 		double weight = cycleBasis.getWeight();
-		Set<GraphPath<Long, Long>> cycles = cycleBasis.getCyclesAsGraphPaths();
+		Set<GraphPath<Integer, Integer>> cycles = cycleBasis.getCyclesAsGraphPaths();
 
 		if (weightRes.isNonNull()) {
 			weightRes.write(weight);
@@ -155,12 +155,12 @@ public class CycleApi {
 			+ "cycles_fundamental_basis_exec_stack_bfs", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executeStackBFSFundamental(IsolateThread thread, ObjectHandle graphHandle,
 			CDoublePointer weightRes, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
 
-		CycleBasisAlgorithm<Long, Long> alg = new StackBFSFundamentalCycleBasis<>(g);
-		CycleBasis<Long, Long> cycleBasis = alg.getCycleBasis();
+		CycleBasisAlgorithm<Integer, Integer> alg = new StackBFSFundamentalCycleBasis<>(g);
+		CycleBasis<Integer, Integer> cycleBasis = alg.getCycleBasis();
 		double weight = cycleBasis.getWeight();
-		Set<GraphPath<Long, Long>> cycles = cycleBasis.getCyclesAsGraphPaths();
+		Set<GraphPath<Integer, Integer>> cycles = cycleBasis.getCyclesAsGraphPaths();
 
 		if (weightRes.isNonNull()) {
 			weightRes.write(weight);
@@ -175,12 +175,12 @@ public class CycleApi {
 			+ "cycles_fundamental_basis_exec_paton", exceptionHandler = StatusReturnExceptionHandler.class)
 	public static int executePatonFundamental(IsolateThread thread, ObjectHandle graphHandle,
 			CDoublePointer weightRes, WordPointer res) {
-		Graph<Long, Long> g = globalHandles.get(graphHandle);
+		Graph<Integer, Integer> g = globalHandles.get(graphHandle);
 
-		CycleBasisAlgorithm<Long, Long> alg = new PatonCycleBase<>(g);
-		CycleBasis<Long, Long> cycleBasis = alg.getCycleBasis();
+		CycleBasisAlgorithm<Integer, Integer> alg = new PatonCycleBase<>(g);
+		CycleBasis<Integer, Integer> cycleBasis = alg.getCycleBasis();
 		double weight = cycleBasis.getWeight();
-		Set<GraphPath<Long, Long>> cycles = cycleBasis.getCyclesAsGraphPaths();
+		Set<GraphPath<Integer, Integer>> cycles = cycleBasis.getCyclesAsGraphPaths();
 
 		if (weightRes.isNonNull()) {
 			weightRes.write(weight);
